refactor: extract shared event test helpers; fix 3d no-pointer_down positive assertion
